初步引入 clickhouse+1

+++ models/clickhouse/v1/manager/manager.go
@@ -1,7 +1,7 @@
 package manager
 
 type Manager struct {
-	ID      uint8  `json:"id"`
+	ID      uint64 `json:"id"`
 	Name    string `json:"name"`
 	Address string `json:"address"`
 	// Phone     string          `json:"phone"`
